refactor(APIM-123): separate AcbsFacilityGuaranteeService tests
